# perf(invariant): sequentially shrink failed sequence