<!--
Thank you for your Pull Request. Please provide a description above and review
the requirements below.

Bug fixes and new features should include tests.
-->

## Motivation

- shrinking call sequences with len > 64 it's not efficient and likely to take really long time without satisfactory result
- atm if call sequences are < 64 we create all possible combinations and test them in parallel, which leads to good results/proper shrinkage
https://github.com/foundry-rs/foundry/blob/19871fcde773659568a141f0755dc8658f117536/crates/evm/evm/src/executors/invariant/error.rs#L328-L340
- if call sequence is > 64 then random samples from original sequence are created and tested in paralel. The number of samples constructed are run at the same time are by default 2^18 / 3 = 87381 (process repeats 3 time). From testing this approach fails to handle even call sequences with a len of 100.
https://github.com/foundry-rs/foundry/blob/19871fcde773659568a141f0755dc8658f117536/crates/evm/evm/src/executors/invariant/error.rs#L341-L359
- while testing different projects it was also noticed that (~~only in some conditions~~ when running with `fail-on-revert=true`) the failed scenario doesn't get reproduced when we try shrinking sequence (~~to investigate why~~ due to the fact we're not checking revert for each call when replaying - this will be addressed in a follow up PR) - in such scenario test hangs. Proposed solution mitigates this issue as it starts checking and reproducing failure from original sequence - if it cannot be replayed then it considers original sequence as the optimized sequence and exits. 

<!--
Explain the context and why you're making that change. What is the problem
you're trying to solve? In some cases there is not a problem and this can be
thought of as being the motivation for your change.
-->

## Solution

<!--
Summarize the solution and provide any necessary context needed to understand
the code change.
-->
- proposed solution take into account that state is composed from calls top down, so the most efficient way to shrink it is to start removing calls from top one by one and test candidate sequence
- if candidate sequence fails then the removed call doesn't affect the failure sequence, so we can move on to next
- if candidate sequence doesn't fail then it is a prereq of failure sequence, so we add back and move to next
- proposed solution is inspired from proptest shrinking basics and use simplify / complicate (even `CallSequenceShrinker` is not a strategy itself) https://altsysrq.github.io/proptest-book/proptest/tutorial/shrinking-basics.html
- from initial testing there is no penalty on shrinking calls under 64 either, whereas calls up to 5000 depth were tested and shrinked in reasonable time 

## Performance tests

* Tests were performed using commit https://github.com/foundry-rs/foundry/commit/008922d5165c764859bc540d7298045eebf5bc60 and current PR (https://github.com/foundry-rs/foundry/pull/7756/commits/9c94eeddc857fbaf00f35e001d0a1a33aa51e332)
* Following test was used (to be able to control number of calls for failure)
```Solidity
contract ShrinkBigSequence {
    uint256 cond;

    function work(uint256 x) public {
        if (x % 2 != 0 && x < 9000) {
            cond++;
        }
    }

    function dummy(uint256 x) public {
        // NOP, just for adding noise in failing call sequence
        require(1 > 2, "NOP");
    }

    function checkCond() public view {
        require(cond < 10, "condition met");
    }
}

contract ShrinkBigSequenceTest is Test {
    ShrinkBigSequence target;

    function setUp() public {
        target = new ShrinkBigSequence();
    }

    /// forge-config: default.invariant.runs = 1
    /// forge-config: default.invariant.depth = 500
    function invariant_shrink_big_sequence() public view {
        target.checkCond();
    }
}
```
* `require(cond < 10, "condition met")` allows setting number of calls for failing scenario (e.g. 10)
* `work` function increments count only in certain conditions, `dummy` function is a NOP, both used to add more noise in failing sequence to be shrinked

#### Fail in 10 steps
`require(cond < 10, "condition met")`  and `/// forge-config: default.invariant.depth = 500`

| Original sequence  | Shrinked to |  Time |  | Original sequence  | Shrinked to |  Time |
| ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- |
| 143 | 10  | 14.49s  | | 119 | 10  | 131ms  |
| 147 | 10 | 15.02s | | 153  | 10 | 206ms |
| 93 | 10 | 10.57s | | 121  | 10 | 142ms |

#### Fail in 20 steps
`require(cond < 20, "condition met")`  and `/// forge-config: default.invariant.depth = 500`

| Original sequence  | Shrinked to |  Time |  | Original sequence  | Shrinked to |  Time |
| ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- |
| 247 | 20  | 67.27s  | | 245 | 20  | 508ms  |
| 277 | 20 | 87.49s | | 260  | 20 | 608ms |
| 252 | 20 | 74.42s | | 177  | 20 | 306ms |

#### Fail in 60 steps
`require(cond < 60, "condition met")`  and `/// forge-config: default.invariant.depth = 1000`

| Original sequence  | Shrinked to |  Time |  | Original sequence  | Shrinked to |  Time |
| ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- |
| 693 | 198  | 469.69s  | | 682 | 60  | 3.96s  |
| - | - | - | | 610  | 60 | 3.23s |
| - | - | - | | 506  | 60 | 2.25s |

#### Fail in 200 steps
`require(cond < 200, "condition met")`  and `/// forge-config: default.invariant.depth = 2000`
Tests with master were stopped after 10 minutes, collected results only for tests with PR

| Original sequence  | Shrinked to |  Time |
| ------------- | ------------- | ------------- |
| 1976 | 200  | 32.88s  |
| 1951 | 200 | 30.91s |
| 1898 | 200 | 29.91s |

#### Fail in 500 steps
`require(cond < 500, "condition met")`  and `/// forge-config: default.invariant.depth = 5000`
Tests with PR only

| Original sequence  | Shrinked to |  Time |
| ------------- | ------------- | ------------- |
| 4589 | 500  | 180.51s  |
| 4931 | 500 | 206.47s |
| 4970 | 500 | 215.45s |

CC @klkvr @mds1